requirements.txt
PySide6>=6.6
mss>=9.0
numpy>=1.26
Pillow>=10.2
scikit-image>=0.22
requests>=2.32
jsonschema>=4.22
pyautogui>=0.9
pynput>=1.7
keyring>=25.2
pytest>=8.2
